automatically backed up by learn

+++ lib/bachelor.rb
@@ -74,7 +74,7 @@ def get_average_age_for_season(data, season)
         data_in_season.each do |data_type|
           data_type.each do |age, value|
             if age == "age"
-              average_age+=value.to_i
+              average_age+=value.to_f 
               counter +=1 
          end
        end
